feat: refactoring soap xml encoding
