+++ src/main/java/it/pagopa/transactions/client/NodeForPspClient.java
@@ -1,18 +1,17 @@
 package it.pagopa.transactions.client;
 
-import java.math.BigDecimal;
-
 import javax.xml.bind.JAXBElement;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
+import org.springframework.http.MediaType;
 import org.springframework.stereotype.Component;
 import org.springframework.web.reactive.function.client.WebClient;
 import org.springframework.web.server.ResponseStatusException;
 
-import it.pagopa.nodeforpsp.ActivatePaymentNoticeReq;
-import it.pagopa.nodeforpsp.ActivatePaymentNoticeRes;
-import it.pagopa.transactions.utils.soap.SoapEnvelopeRequest;
+import it.pagopa.transactions.model.ActivatePaymentNoticeReq;
+import it.pagopa.transactions.model.ActivatePaymentNoticeRes;
+import it.pagopa.transactions.model.SoapEnvelope;
 import lombok.extern.slf4j.Slf4j;
 import reactor.core.publisher.Mono;
 
@@ -25,36 +24,21 @@ public class NodeForPspClient {
 
 	public Mono<ActivatePaymentNoticeRes> activatePaymentNotice(JAXBElement<ActivatePaymentNoticeReq> request) {
 
-		return nodoWebClient.post().body(Mono.just(new SoapEnvelopeRequest("", request)), SoapEnvelopeRequest.class)
+		return nodoWebClient.post().header("Content-Type", MediaType.TEXT_XML_VALUE)
+				.body(Mono.just(new SoapEnvelope("", request)), SoapEnvelope.class)
 				.retrieve()
 				.onStatus(HttpStatus::isError,
 						clientResponse -> clientResponse.bodyToMono(String.class)
 								.flatMap(errorResponseBody -> Mono.error(
 										new ResponseStatusException(clientResponse.statusCode(), errorResponseBody))))
-				.bodyToMono(String.class)
-				.map(response -> {
-					// TODO temporary solution, waiting for Jaxb2SoapDecoder.java
-					ActivatePaymentNoticeRes activatePaymentNoticeRes = new ActivatePaymentNoticeRes();
-					String paymentToken = response
-							.substring(response.indexOf("<paymentToken>") + "<paymentToken>".length());
-					activatePaymentNoticeRes
-							.setPaymentToken(paymentToken.substring(0, paymentToken.indexOf("</paymentToken>")));
-					String totalAmount = response
-							.substring(response.indexOf("<totalAmount>") + "<totalAmount>".length());
-					activatePaymentNoticeRes.setTotalAmount(
-							new BigDecimal(totalAmount.substring(0, totalAmount.indexOf("</totalAmount>"))));
-					String paymentDescription = response
-							.substring(response.indexOf("<paymentDescription>") + "<paymentDescription>".length());
-					activatePaymentNoticeRes.setPaymentDescription(
-							paymentDescription.substring(0, paymentDescription.indexOf("</paymentDescription>")));
-					return activatePaymentNoticeRes;
-				})
+				.bodyToMono(ActivatePaymentNoticeRes.class)
 				.doOnSuccess((ActivatePaymentNoticeRes paymentActivedDetail) -> {
-					log.debug("Payment activated with paymentToken {}", paymentActivedDetail.getPaymentToken());
+					log.debug("Payment activated with paymentToken {}",
+							new Object[] { paymentActivedDetail.getPaymentToken() });
 				}).doOnError(ResponseStatusException.class, error -> {
-					log.error("error : {}", error);
+					log.error("error : {}", new Object[] { error });
 				}).doOnError(Exception.class, (Exception error) -> {
-					log.error("error : {}", error);
+					log.error("error : {}", new Object[] { error });
 				});
 	}
 }
