+++ src/main/java/it/pagopa/transactions/configurations/WebClientsConfig.java
@@ -12,7 +12,8 @@ import org.springframework.web.reactive.function.client.WebClient;
 import io.netty.channel.ChannelOption;
 import io.netty.handler.timeout.ReadTimeoutHandler;
 import it.pagopa.ecommerce.sessions.v1.ApiClient;
-import it.pagopa.nodeforpsp.ObjectFactory;
+import it.pagopa.transactions.model.ObjectFactory;
+import it.pagopa.transactions.utils.soap.Jaxb2SoapDecoder;
 import it.pagopa.transactions.utils.soap.Jaxb2SoapEncoder;
 import reactor.netty.http.client.HttpClient;
 
@@ -29,12 +30,12 @@ public class WebClientsConfig {
                         .addHandlerLast(new ReadTimeoutHandler(nodoReadTimeout, TimeUnit.MILLISECONDS)));
 
         ExchangeStrategies exchangeStrategies = ExchangeStrategies.builder().codecs(clientCodecConfigurer -> {
+            clientCodecConfigurer.registerDefaults(false);
+            clientCodecConfigurer.customCodecs().register(new Jaxb2SoapDecoder());
             clientCodecConfigurer.customCodecs().register(new Jaxb2SoapEncoder());
-            // clientCodecConfigurer.customCodecs().register(new Jaxb2SoapDecoder()); TODO
-            // fix Jaxb2SoapDecoder
         }).build();
 
-        return WebClient.builder().baseUrl(nodoUri).defaultHeader("Content-Type", "text/xml")
+        return WebClient.builder().baseUrl(nodoUri)
                 .clientConnector(new ReactorClientHttpConnector(httpClient)).exchangeStrategies(exchangeStrategies)
                 .build();
     }
