+++ src/main/java/it/pagopa/transactions/utils/soap/Jaxb2SoapDecoder.java
@@ -5,10 +5,11 @@ import org.springframework.core.codec.CodecException;
 import org.springframework.core.codec.DecodingException;
 import org.springframework.core.io.buffer.DataBuffer;
 import org.springframework.core.io.buffer.DataBufferUtils;
+import org.springframework.http.MediaType;
 import org.springframework.http.codec.xml.Jaxb2XmlDecoder;
 import org.springframework.lang.Nullable;
 import org.springframework.util.MimeType;
-import org.springframework.util.xml.StaxUtils;
+import org.springframework.util.MimeTypeUtils;
 import org.springframework.ws.WebServiceMessage;
 import org.springframework.ws.WebServiceMessageFactory;
 import org.springframework.ws.client.core.WebServiceTemplate;
@@ -19,26 +20,25 @@ import javax.xml.bind.JAXBElement;
 import javax.xml.bind.JAXBException;
 import javax.xml.bind.UnmarshalException;
 import javax.xml.bind.Unmarshaller;
-import javax.xml.stream.XMLInputFactory;
 import javax.xml.stream.XMLStreamException;
+
 import java.util.Map;
 
 public class Jaxb2SoapDecoder extends Jaxb2XmlDecoder {
 
-    private static final XMLInputFactory inputFactory = StaxUtils.createDefensiveInputFactory();
-
     private final JaxbContextContainer jaxbContexts = new JaxbContextContainer();
 
     public Jaxb2SoapDecoder() {
-        super();
+        super(MimeTypeUtils.APPLICATION_XML, MimeTypeUtils.TEXT_XML, new MediaType("application", "*+xml"),
+                new MediaType("text", "*"));
     }
 
     @Override
-    @SuppressWarnings({ "rawtypes", "unchecked", "cast" })
     public Object decode(DataBuffer dataBuffer, ResolvableType targetType, @Nullable MimeType mimeType,
             @Nullable Map<String, Object> hints) throws DecodingException {
 
         try {
+
             DefaultStrategiesHelper helper = new DefaultStrategiesHelper(WebServiceTemplate.class);
             WebServiceMessageFactory messageFactory = helper.getDefaultStrategy(WebServiceMessageFactory.class);
             WebServiceMessage message = messageFactory.createWebServiceMessage(dataBuffer.asInputStream());
@@ -53,7 +53,8 @@ public class Jaxb2SoapDecoder extends Jaxb2XmlDecoder {
 
     private Object unmarshal(WebServiceMessage message, Class<?> outputClass) {
         try {
-            Unmarshaller unmarshaller = initUnmarshaller(outputClass);
+
+            Unmarshaller unmarshaller = getUnmarshaller();
             JAXBElement<?> jaxbElement = unmarshaller.unmarshal(message.getPayloadSource(), outputClass);
             return jaxbElement.getValue();
         } catch (UnmarshalException ex) {
@@ -63,9 +64,9 @@ public class Jaxb2SoapDecoder extends Jaxb2XmlDecoder {
         }
     }
 
-    private Unmarshaller initUnmarshaller(Class<?> outputClass) throws CodecException, JAXBException {
-        Unmarshaller unmarshaller = this.jaxbContexts.createUnmarshaller(outputClass);
-        return getUnmarshallerProcessor().apply(unmarshaller);
+    private Unmarshaller getUnmarshaller() throws JAXBException {
+
+        return jaxbContexts.createUnmarshaller();
     }
 
 }
