--- src/main/java/it/pagopa/transactions/utils/soap/SoapEnvelopeRequest.java
@@ -1,27 +1,0 @@
-package it.pagopa.transactions.utils.soap;
-
-import javax.xml.bind.annotation.XmlRootElement;
-
-@XmlRootElement
-public class SoapEnvelopeRequest {
-
-    private String headerContent;
-    private Object body;
-
-    public SoapEnvelopeRequest(String headerContent, Object body) {
-        this.headerContent = headerContent;
-        this.body = body;
-    }
-
-    public SoapEnvelopeRequest(Object body) {
-        this.body = body;
-    }
-
-    public String getHeaderContent(){
-        return headerContent;
-    }
-
-    public Object getBody() {
-        return body;
-    }
-}
